Add a deck of cards
While this commit seems simple enough, a lot of work ultimately went into it, including some of my favourite code. Here's a quick list:

- Add entitlements file (So that the app is appropriately sandboxed)
- Rename the application to 'blackjack', while keeping the project named 'Blackjack'
- Rename 'GameScene' to 'Game' (2 instances)
- Add a Card class
    - This is an extension of `SKSpriteNode`, which is used to display image sprites. This may change in a future update.
    - The class file includes Suit and Rank, two enums which hold suit and rank data
        - Suit is either Spades, Hearts, Diamonds or Clubs, and can return the symbols for each type
        - Rank is the rank of a card, and can return that in a textual value (numerical will need to be determined later)
- Build a deck of cards by iterating through Suit and Rank's `rawValue`s
- Shuffle the deck of cards using GameplayKit
- Add an algorithm for shuffling of my own creation to fulfill programming requirements

The piece of code that I'm incredibly proud of is the whole Card class (it's really clean and organised), especially Rank.value().
This function prints 'A', 'J', 'Q', or 'K' for special cards, and the value for individual cards—by converting the rawValue.
This way, we cleverly minimise the need for a large amount of `case`s in our `switch`. Head into Card.swift and read it, seriously.