Display cards
This commit allows us to display each card, in a way that would make sense to players of Blackjack (or any other card game, for that matter). If you were to download this part of the project, by itself, it could make a very good model for almost any card game.

When we display cards, we use the Unicode card set (https://en.wikipedia.org/wiki/Playing_cards_in_Unicode#Block), which is a pretty standard deck for this sort of thing. Instead of writing a really big `switch` statement, though, I realised we could use the inherent property of Unicode in that it's _ordered_. Not only ordered, but helpfully split into rows on a table. For cards, which is nice, each row represents a different suit, so we can pretty easily convert suits and ranks into a number which will let us build a card into the game as we need.

Once that was done, I had to give the cards backgrounds. While I could use a rounded rectangle to match the card, we can avoid a performance hit by using an ordinary rectangle, and placing it inside the border of the card (there's a very small spot there where we can put a rectangle without there being any gaps or overflow).

Also of importance was removing the line in `AppDelegate.swift` which helps with performance by ignoring the sibling order. When we do this, the cards and their backgrounds often become separated by other elements—we need to render cards _and_ their backgrounds together, so I disabled it.

Finally, the routine in `update()` adds the 52 decks I've specified to the game, in their random positions. This will go when we're done, but it's a nice placeholder _and_ performance test. It holds up pretty well.