Change card style
In line with my objectives, I'm trying to make Blackjack a little more efficient. For a blackjack player, the traditional card display doesn't really matter beyond the number, and if you don't card-count the suit doesn't even matter. So why bother displaying some weird huge card when we can compress it to a smaller, nicer shape. Although the ideal shape here is probably a triangle, it doesn't tesselate well, so I chose a parallelogram instead. A rectangle shape was also cut, because it wasn't interesting enough.

The slanted sides of the parallelogram are much sleeker and nicer to look at, and they can tesselate well as necessary. They also don't need to overlap on top of other cards, which is nice.