Add players
This commit adds a Player class, which computerised and human players should eventually subclass. Each player has a hand which cards are dealt to, and also $500 starting cash. I was feeling generous.

We can also appropriately calculate the player's score now, which is nice. Upon calling `Player.score()`, the caller will receive a dictionary with appropriate numbers for Aces (depending on how they are scored). If someone is reading this, I'd really suggest reading the inline documentation on this, because I'm pretty proud of it. Consider it part of the 'Build Log' component. I've been writing a lot of 'good' code recently, which is nice.

Beyond that, we also have a `hit()` function for players who want more cards. It's modular, so we also use it to initially deal two cards. Cool!

I also cleaned up a few things around that weren't being used.