Display cards and let player hit
This commit includes a mechanism for displaying an infinite number of cards in the player's hand, and allows them to add cards to their hand.

I made the conscious choice this time to not allow for mouse control. If my game is for Blackjack enthusiasts, then they're unlikely to want to use a mouse when keyboard controls are much more accessible and easier to use.