Implement MVC to make room for AI
I'll be implementing AI soon, but it's a lot easier on resources if we implement MVC first. So I spent most of today splitting up the framework for the view controller (GameScene) from the framework for the model (Game). There's also a bunch of other code in here to get ready for AI, once I figure out why it's bugging out. Woe is me.