Add dealer and fix hitting loop
It turns out AI might be unreasonable, so I re-implemented shouldAIHit() with my own algorithm to determine when to hit. I also added a dealer to the list of players, who holds the hand we all play against. Although, there are some dealer-y rules about visible cards I'll need to implement soon. This project seems simple but it's actually really long and burdensome and annoying. -_-