Finalise the game loop
As of this commit, the game loop works properly. We can now test when the game is over, and reset it appropriately. This way, you can keep holding down 'H' to hit, and see where it gets you.
Past that, I also added some extra money logic so that the game appropriately invests money into the game and receives the correct returns on the investment.