Double down, and customise AI/dealer
This update enables basic customisation of the AI players and the dealer. For the AI players, we implement an 'Ace Risk'—the chance of them continuing to hit if they have an ace in their hand. The player can adjust this in 10% increments using the '+' and '-' keys on their keyboard.
The dealer can be toggled between hitting or standing on a soft 17, using the 'T' key.
The player can also double down, which is a hit followed by a stand, by pressing 'D'. I've chosen to presume that the player knows what this means, in line with the problem definition and solution constraints.