Properly double bets
I had a bit of a blank when implementing doubling. I didn't understand why you'd want to hit then stand if just doing them gave you more freedom. Turns out I totally forgot why doubling is beneficial.
So this commit properly doubles the player's bets. It also leaves room for the bet to be changed, but right now that's not a priority. This could very well be the last commit.