Defeat the house edge
I tasked myself today with finding a spot where my game really stands out from the others. To do this, I figured my time would be better spent on fixing the AI instead of implementing different betting (which I can do but isn't really part of the marking criteria).

The first thing I realised is that a card counter could be easily implemented because the cards left in the deck are the cards that aren't visible (think about it). So, I iterated through all of the cards left in the deck. For each, I determined the conditions if it were added, and added it to a 'guessFactor'. If the player goes bust when adding the card, add 0. If they don't, add 1. Then we'd divide this number by the number of cards, and determine whether the move was favourable or not by rounding.

This didn't work. I wasn't regularly hitting 1000 rounds like I would normally.

I tried re-implementing a chance of doubling, and the 'ace risk', but they appeared to make this worse. Eventually I realised there was an issue with the ace risk—I'd determine when to _hit_ instead of when to stand, and then move onward to the rest of the code. The problem was that the rest of the code would suggest a hit. So I reversed this logic and started to improve the number of rounds I'd hit. But still not anywhere close to 1000.

Finally, I realised that I could just do what I was doing before. We just calculate the average total score once each card is added, and if it's less than 21 we hit. With this, plus an improved doubling and ace risk algorithm, we finally reached 1000 rounds again. I left the program running, and 10 minutes later, I was on 30000! And the players were still gaining money!

While this doesn't happen every time (because $500 isn't a good enough safety net), players that survive the first 1000 or so rounds will usually go on to keep winning money, indefinitely. This is awesome!